third_party/blink/web_tests/fast/workers/chromium/worker-console-log-expected.txt

CONSOLE MESSAGE: log
CONSOLE MESSAGE: function
CONSOLE MESSAGE: function log() { [native code] }
CONSOLE ERROR: error
CONSOLE WARNING: warn
CONSOLE MESSAGE: info
CONSOLE DEBUG: debug
CONSOLE DEBUG: count: 1
CONSOLE ERROR: console.assert
This tests that 'console.log' and friends function correctly from workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
